Implementado el guardado de monedas/Constructora de PescaState en el cpp

+++ captain_brinetooth/captain_brinetooth/states/PasueState.h
@@ -8,7 +8,7 @@ class Entity;
 struct infoPartida
 {
 	bool abilities[6];
-	float points;
+	int points;
 	//etc..
 };
 static infoPartida partida;
